check

+++ frontend/src/components/WorkoutDetails.jsx
@@ -10,6 +10,10 @@ function WorkoutDetails({workout}) {
   const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return 
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
       headers: {
